Changed bit parsing
I realized that the QA bit parsing code here was incorrect for the output we had described in the paper, and thus have replaced it with a set function directly from the MCD12Q2 user guide (link added to function). 

Note that this update to the QA does NOT change the data used for the paper because 1) the phenometric data was already best quality and 2) the evi data was filtered to only have best quality regardless. However, I thought it best to have the correct code represented here.